support local code sign
Match-id-4b4451096f2c43cb055ed7de351d82b5926a9408

+++ interfaces/innerkits/code_sign_utils/include/code_sign_utils.h
@@ -46,7 +46,7 @@ public:
      * @param len length of signature data
      * @return err code, see err_code.h
      */
-    static int32_t EnforceCodeSignForFile(const std::string &path, const uint8_t *signature, const size_t len);
+    static int32_t EnforceCodeSignForFile(const std::string &path, const uint8_t *signature, const uint32_t len);
     /**
      * @brief Enforce code signature for file with signature
      * @param path file path
